requirements.txt
altair
numpy
pandas
pydeck
streamlit
statsmodel
matplotlib
sklearn
openpyxl